add inspire_id and resource_id as URL parameters
