+++ src/core/hepdata.ts
@@ -1,6 +1,6 @@
 import { useStoreIDStore } from 'src/stores/storeid';
 import { useWorkspaceStore } from 'src/stores/workspace';
-import type { IAnalysis, IHEPdataentry } from 'src/interfaces';
+import type { IAnalysis, IHEPdataentry, IWorkspace } from 'src/interfaces';
 import { Notify } from 'quasar';
 
 export async function check_workspaces_on_HEPdata(
@@ -55,10 +55,36 @@ export async function check_workspaces_on_HEPdata(
 }
 
 export async function load_workspaces_from_HEPdata(analyses: IAnalysis[]): Promise<void> {
-  const store_id_store = useStoreIDStore();
   for (const analysis of analyses) {
-    const id = store_id_store.add_store_with_id();
-    const workspace_store = useWorkspaceStore(id)();
-    workspace_store.load_workspace_from_HEPdata(analysis);
+    load_workspace_from_HEPdata(analysis);
   }
+}
+
+export async function load_workspace_from_HEPdata_resource(record_id: string): Promise<void> {
+  const analysis: IAnalysis = {name: 'Resource ID: ' + record_id, url: 'https://www.hepdata.net/record/resource/'+record_id+'?landing_page=true'}
+  load_workspace_from_HEPdata(analysis);
+}
+
+export async function load_workspace_from_HEPdata(analysis: IAnalysis): Promise<void> {
+  const store_id_store = useStoreIDStore();
+  const id = store_id_store.add_store_with_id();
+  const workspace_store = useWorkspaceStore(id)();
+  const response = await (
+    await fetch(analysis.url.replace('landing_page=true', 'format=json'))
+  ).json()
+  let workspace = {} as IWorkspace
+  if (response.file_contents === 'Large text file') {
+    workspace = await (
+      await fetch(
+        analysis.url.replace('landing_page=true', 'view=true')
+      )
+    ).json();
+  }
+  else {
+    workspace = JSON.parse(response.file_contents);
+  }
+  workspace_store.workspace = workspace;
+  workspace_store.name = analysis.name;
+  workspace_store.loading = false;
+  workspace_store.create_channel_stores();
 }
